Add a list of queries to the top of the query debugger.
This list aggregates the query runtime and number of calls to the
query. This is much more useful, imo, than the text view below.

Note that this patch also disables the text view by default, as it
can easily lead to out-of-memory issues. Blindly appending lots
of text is bound to fail, sooner or later.

The context menu is replaced by a tool button. This makes it also
possible to search in the text edit, and to copy text from it.
Much more useful.

Add average duration and aggregated totals row.

REVIEW: 121307