Registers on DBus with an instance-specific name.
akonadiconsole allows to choose which akonadi instance to monitor, but then
for trackers the DBus name was always "org.kde.akonadiconsole", so it was
being polluted by jobs from the main instance.
Fixed now, e.g. it registers as "org.kde.akonadiconsole-testrunner-20593".
The main instance is still "org.kde.akonadiconsole" so compat is kept.