Check if collection contains items without RID before clearing cache
Items without RID were likely never uploaded to the server and exist
only in Akonadi, so clearing the cache means they will be gone for
good. To make sure users are aware of it (since we recommend using
this feature as a troubleshooting step) we now check for such items
and show a big warning.

Also added a bunch of TODOs on how to improve the clear cache feature.