Enable the Save Button if an Attribute is Deleted
Summary:
Regardless of whether the model's row count is 0 or not, if a
row is deleted, the user has to be able to save the modification.

Test Plan:
Create an attribute and save it. Then delete it and try to save it. Before this patch the last save was not
possible.

Reviewers: #kde_pim, dvratil

Reviewed By: #kde_pim, dvratil

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D8981