Scrollable Area for Browser Internals
Summary:
The contents of the "Internal" tab in the "Browser" tab are too long for
a 1366x768 screen which makes the lower buttons inaccessible.

Test Plan:
Try to resize the Akonadi Console window to a very small size. Before this patch it would not resize beyond certain
point.

Reviewers: #kde_pim, mlaurent

Reviewed By: #kde_pim, mlaurent

Subscribers: mlaurent

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D9024