Akonadiconsole: support incidences serialized in the binary format
Summary:
This experiment shows, however, that the "raw payload" really isn't the
actual raw payload from the DB, but rather the deserialized payload,
serialized again, i.e. it's "raw" only if the serializer is
actually symmetrical...

For this reason I'm wondering if the "converted from binary" comment is
a good idea after all, it makes it sound like the raw payload in the DB
is in the binary format, even when it's not.

Test Plan:
clicking on events in akonadiconsole, with KCALCORE_BINARY_SERIALIZER=1,
shows "converted from binary" for all events. Without this patch, it showed binary
garbage for all events.

Reviewers: dvratil, smartins

Reviewed By: dvratil

Subscribers: #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D12461